space mod is fixed! and comments added
When typing fast, space would not be sent because the key released logic
was resetting key_check. Now key_check is only reset when a key is
pressed which fixes everything. I also reversed the true/false logic
here because I think it makes more sense this way, and I should probably
find a better name for the variable too. Set the var to bool because
that's what it is. Saves a byte of memory.